Add access control components
Resolves #24.

Introduces the `RequireAuthentication` component, which restricts access
to authenticated (with Firebase) users (or alternatively,
unauthenticated users, if the `reverse` prop is set), and navigates to
`/login` otherwise.

Introduces the `RequireSpotifyToken` component, which restricts access
to users with a non-expired token in the store, and navigates to
`/settings` otherwise.

Introduces the `WaitingView` component, which is shown by the other
components when the required data is not available yet, as well as by
the callback page when updating info.

Updates the routes in `index.js` to use these new components, and
updates pages to remove previous checks and work with the new
components.